+++ src/App.js
@@ -1,33 +1,5 @@
-import React, { Component } from 'react';
-import './App.css';
-
-class Counter extends Component {
-  constructor() {
-    super();
-
-    this.state = {
-      counter: 0,
-    };
-  }
-
-  onClickButton = () => {
-    this.setState(prevState => ({
-      counter: prevState.counter + 1,
-    }));
-  }
-
-  render() {
-    const { onClickButton, state: { counter } } = this;
-    console.log(this);
-
-    return (
-      <>
-        <button type="button" onClick={onClickButton}>Click to add</button>
-        <div>{counter}</div>
-      </>
-    );
-  }
-}
+import React from 'react';
+import Counter from './components/Counter/Counter';
 
 function App() {
   return (
